+++ docs/releases.md
@@ -7,6 +7,7 @@ The changes have been tracked starting version 0.1.0
 
 * Fix an issue where segments were not considered in findComponentUsage when they were used in columns
 * adding `RequestCreator` class
+* fix issue when logging object error the output of failed calls.
 
 ## version 0.2.9